requirements.txt
pandas
scikit-learn
matplotlib
imbalanced-learn
tensorflow

# Optional: add any other packages you use (e.g., numpy, seaborn)
